Make TransferRequest timeouts editable with reflection
I am a Relativity engineer, we are using Data Movement library to transfer huge portions of data. We experience failures for slower networks and bigger files. We were able to solve this by overriding those timeouts with reflection, unfortunately this doesn't work on .net core as those properties are marked 'readonly'. Alternatively I could make this properly and expose those parameters to client, although for our usecase this is sufficient.